Added volume adjustment, but it is not yet synced with the start and ends of the sine path.
frequencies has been renamed frequencies2 to avoid clash with core/frequencies.
The crazy (map #(map ... calls have been moved inside fns freqs-to-sines and volume-adjust-sines. I have the feeling that working on lists of lists of frequencies is a bad idea and that that central fn should not be frequencies2, but instead a call to (range threads) and then each map on that can deal with a single column and will also have awareness of where it fits in the grand scheme of things.